取得したデータのキー名を日本語から英語に変換完了
