+++ front/src/App.tsx
@@ -2,6 +2,7 @@ import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import Router from './Router';
 import { Header } from './components/header';
+import { Prefecture } from './components/atoms/Prefecture';
 
 // slice
 import { fetchDailyInfectionAsync } from './features/graphs/dailyInfectionSlice';
@@ -16,6 +17,7 @@ const App = () => {
     <>
       <Header />
       <Router />
+      <Prefecture />
     </>
   );
 };
