+++ front/src/common/Prefecture.tsx
@@ -1,12 +1,4 @@
-import React, { useState } from 'react';
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
-import InputLabel from '@material-ui/core/InputLabel';
-import MenuItem from '@material-ui/core/MenuItem';
-import FormHelperText from '@material-ui/core/FormHelperText';
-import FormControl from '@material-ui/core/FormControl';
-import Select from '@material-ui/core/Select';
-
-const prefectures = [
+export const prefectures = [
   { prefCode: 1, prefName: '北海道' },
   { prefCode: 2, prefName: '青森県' },
   { prefCode: 3, prefName: '岩手県' },
@@ -54,43 +46,3 @@ const prefectures = [
   { prefCode: 46, prefName: '鹿児島県' },
   { prefCode: 47, prefName: '沖縄県' },
 ];
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    formControl: {
-      margin: theme.spacing(1),
-      minWidth: 120,
-    },
-    selectEmpty: {
-      marginTop: theme.spacing(2),
-    },
-  })
-);
-
-export const Prefecture = () => {
-  const classes = useStyles();
-  const [prefecture, setPrefecture] = useState('');
-  const handleChange = (event: React.ChangeEvent<{ value: unknown }>) => {
-    setPrefecture(event.target.value as string);
-  };
-  return (
-    <>
-      <FormControl className={classes.formControl}>
-        <InputLabel id="demo-simple-select-helper-label">都道府県</InputLabel>
-        <Select
-          labelId="demo-simple-select-helper-label"
-          id="demo-simple-select-helper"
-          value={prefecture}
-          onChange={handleChange}
-        >
-          {prefectures.map((prefecture, index) => (
-            <MenuItem key={index} value={String(prefecture.prefCode)}>
-              {prefecture.prefName}
-            </MenuItem>
-          ))}
-        </Select>
-        <FormHelperText>都道府県を選択してください。</FormHelperText>
-      </FormControl>
-    </>
-  );
-};
